Best practices
"Do not set the value to 0, which results in displaying the password expiration warning every time the user logs on." needs to be updated. 
As per my testing the user will not get password expiry notification on every logon if we set the value to 0. User will get notification on last day of password expiry